src/requirements_r1.txt
openai==1.59.6
langchain-openai==0.3.0

chromadb==0.5.23  
#現在:chromadbのLatest version:0.6.2ですが、langchain-chromaは >=0.4.0 かつ <0.6.0
#また以下バージョンは除外されます:0.5.10, 0.5.11, 0.5.12, 0.5.4, 0.5.5, 0.5.7, 0.5.9
langchain-chroma==0.2.0

langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.5
langsmith==0.2.10
streamlit==1.41.1
streamlit-feedback==0.1.3

python-dotenv==1.0.1

pypdf==5.0.1
pycryptodome==3.21.0